renamed folder to EventsDashboard
