+++ src/components/EventsDashboard/EventCard.tsx
